move router to it's own module
